// loan-origination/src/app/feature/BNPL/shared/models/general-parameters-response.model.ts
// Generated by https://quicktype.io

export interface GeneralParameter {
  id:                 number;
  identification:     ActivityContractor[];
  creditUse:          ActivityContractor[];
  feesMonthly:        ActivityContractor[];
  companyActivity:    ActivityContractor[];
  pensionCompany:     ActivityContractor[];
  activityContractor: ActivityContractor[];
  department:         ActivityContractor[];
  city:               ActivityContractor[];
  civilStatus:        ActivityContractor[];
  career:             ActivityContractor[];
  relationship:       ActivityContractor[];
  producType:         ActivityContractor[];
  entity:             ActivityContractor[];
  occupation:         ActivityContractor[];
  assets:             ActivityContractor[];
  notice:             ActivityContractor[];
  profession:         ActivityContractor[];
}

export interface ActivityContractor {
  id:          number;
  description: string;
}