// loan-origination/src/app/core/shared/models/credit-approbe.model.ts
// Generated by https://quicktype.io

export interface CreditApprobe {
  amountFinance:      number;
  fees:               number;
  rateApprobeAnnual:  number;
  rateApprobeMonth:   number;
  fGA:                number;
  platform:           number;
  validationStrategy: string;
  paid: string;
  cuotaBenefits: number;
  feesBenefits: number;
}