// loan-origination/src/app/feature/BNPL/shared/models/payment.model.ts
// Generated by https://quicktype.io
// Generated by https://quicktype.io

export interface Payment {
  amountFinance:      number;
  fees:               number;
  rateApprobeAnnual:  number;
  rateApprobeMonth:   number;
  fGA:                number;
  platform:           number;
  datePaid:           string;
  numCredit:          string;
  minPaid:            string;
  validationStrategy: string;
  cuotaBenefits:     number;
  feesBenefits:      number;
  paid:              string;
}